+++ packages/form-react/src/types/index.ts
@@ -10,6 +10,7 @@ export type {
   SelectFieldProps,
   SelectableFieldProps,
   RadioFieldProps,
+  RatingFieldProps,
   CheckboxGroupFieldProps,
   SwitchFieldProps,
   SearchFieldProps,